backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
httpx==0.27.0
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.35
python-multipart==0.0.9
numpy==1.26.0
Pillow==10.4.0
